Break out details form store file into core segments
