+++ src/components/shared/Entity/Edit/DraftInitializer.tsx
@@ -1,5 +1,6 @@
+import FullPageSpinner from 'components/fullPage/Spinner';
 import useInitializeTaskDraft from 'components/shared/Entity/Edit/useInitializeTaskDraft';
-import { useEffect } from 'react';
+import { useEffect, useState } from 'react';
 import { useFormStateStore_status } from 'stores/FormState/hooks';
 import { FormStatus } from 'stores/FormState/types';
 import { BaseComponentProps } from 'types';
@@ -10,14 +11,20 @@ function DraftInitializer({ children }: BaseComponentProps) {
     // Form State Store
     const formStatus = useFormStateStore_status();
 
+    const [loading, setLoading] = useState<boolean>(true);
+
     useEffect(() => {
         if (formStatus === FormStatus.INIT) {
-            void initializeTaskDraft();
+            void initializeTaskDraft(setLoading);
         }
-    }, [initializeTaskDraft, formStatus]);
+    }, [initializeTaskDraft, setLoading, formStatus]);
 
-    // eslint-disable-next-line react/jsx-no-useless-fragment
-    return <>{children}</>;
+    if (loading) {
+        return <FullPageSpinner />;
+    } else {
+        // eslint-disable-next-line react/jsx-no-useless-fragment
+        return <>{children}</>;
+    }
 }
 
 export default DraftInitializer;
