+++ src/stores/DetailsForm/types.ts
@@ -1,0 +1,41 @@
+import { JsonFormsCore } from '@jsonforms/core';
+
+export interface Details extends Pick<JsonFormsCore, 'data' | 'errors'> {
+    data: {
+        description?: string;
+        entityName: string;
+        connectorImage: {
+            id: string;
+            iconPath: string;
+            imagePath: string;
+            connectorId: string;
+        };
+    };
+}
+
+export interface DetailsFormState {
+    // Form Data
+    details: Details;
+    setDetails: (details: Details) => void;
+    setDetails_connector: (
+        connector: Details['data']['connectorImage']
+    ) => void;
+
+    detailsFormErrorsExist: boolean;
+
+    // Connectors
+    connectors: { [key: string]: any }[];
+    setConnectors: (val: DetailsFormState['connectors']) => void;
+
+    // Misc.
+    draftedEntityName: string;
+    setDraftedEntityName: (
+        value: DetailsFormState['draftedEntityName']
+    ) => void;
+
+    entityNameChanged: boolean;
+    setEntityNameChanged: (value: string) => void;
+
+    stateChanged: () => boolean;
+    resetState: () => void;
+}
